# api/openapi.yaml
openapi: 3.1.0
info:
  title: Prive Exchange API
  version: 0.1.0-phase1
  description: >
    Phase-1 (MVP) REST surface for Prive Exchange — a hybrid carbon-credit exchange.
    Formal companion to docs/03-api-contracts.md. Money values are decimal STRINGS, never
    numbers. All money-moving POSTs require an Idempotency-Key header. WebSocket and GraphQL
    surfaces are specified separately (see docs/03).
  contact:
    name: Prive Exchange Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.priveexchange.com/v1
    description: Production
  - url: https://api.staging.priveexchange.com/v1
    description: Staging (Polygon Amoy testnet)
  - url: http://localhost:8080/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /orders:
    post:
      tags: [Trading]
      summary: Place an order
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlaceOrderRequest' }
      responses:
        '201':
          description: Order accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Market halted or self-trade prevented
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    get:
      tags: [Trading]
      summary: List own orders
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/OrderStatus' }
        - name: market
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Order' }
                  next_cursor: { type: [string, 'null'] }
...
  /retirements:
    post:
      tags: [Retirement]
      summary: Retire (burn) credits for offsetting
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [batch_id, quantity]
              properties:
                batch_id: { type: string, format: uuid }
                quantity: { $ref: '#/components/schemas/Decimal' }
                beneficiary: { type: string, example: 'Acme Steel Ltd FY2026' }
      responses:
        '202':
          description: Retirement accepted (on-chain confirmation follows)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Retirement' }
    get:
      tags: [Retirement]
      summary: Own retirement history
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Retirement' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshAuth:
      type: http
      scheme: bearer
      description: Refresh token (rotated on use)
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Required on all money-moving requests; replays return the original result.
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
...
    Forbidden:
      description: Authenticated but not permitted (RBAC / tier / flags)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    UnprocessableEntity:
      description: Validation failed (e.g. insufficient funds, tick/lot violation)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Decimal:
      type: string
      pattern: '^-?[0-9]+(\.[0-9]+)?$'
      description: Decimal number as a string to avoid float loss.
      example: '12.500000'
    Problem:
      type: object
      description: RFC-7807 problem+json.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        trace_id: { type: string }
...
    OrderSide:
      type: string
      enum: [buy, sell]
    OrderType:
      type: string
      enum: [market, limit, stop_limit, stop_market]
      description: Phase-1 set. oco/iceberg/twac/recurring are Phase 2 (501 until enabled).
    Tif:
      type: string
      enum: [gtc, ioc, fok]
    OrderStatus:
      type: string
      enum: [open, partial, filled, cancelled, rejected]
    PlaceOrderRequest:
      type: object
      required: [market, side, type, quantity]
      properties:
        market: { type: string, example: 'AMZN-RF25/USDT' }
        side: { $ref: '#/components/schemas/OrderSide' }
        type: { $ref: '#/components/schemas/OrderType' }
        tif: { $ref: '#/components/schemas/Tif' }
        price: { $ref: '#/components/schemas/Decimal' }
        stop_price: { $ref: '#/components/schemas/Decimal' }
        quantity: { $ref: '#/components/schemas/Decimal' }
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        market: { type: string }
        side: { $ref: '#/components/schemas/OrderSide' }
        type: { $ref: '#/components/schemas/OrderType' }
        status: { $ref: '#/components/schemas/OrderStatus' }
        price: { $ref: '#/components/schemas/Decimal' }
        quantity: { $ref: '#/components/schemas/Decimal' }
        filled_qty: { $ref: '#/components/schemas/Decimal' }
        avg_fill_price: { type: [string, 'null'] }
        hold_id: { type: [string, 'null'], format: uuid }
        fee_estimate: { $ref: '#/components/schemas/Decimal' }
        co2_tonnes: { $ref: '#/components/schemas/Decimal' }
        reject_reason: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
...
    Retirement:
      type: object
      properties:
        id: { type: string, format: uuid }
        batch_id: { type: string, format: uuid }
        quantity: { $ref: '#/components/schemas/Decimal' }
        status: { type: string, enum: [pending, confirmed] }
        certificate_token_id: { type: [string, 'null'] }
        certificate_ipfs_cid: { type: [string, 'null'] }
        tx_hash: { type: [string, 'null'] }
        explorer_url: { type: [string, 'null'], format: uri }
...